spurious "(uint)" cast which was causing compilation warnings removed
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@48 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/checklst.cpp
@@ -292,7 +292,7 @@ void wxCheckListBox::OnChar(wxKeyEvent& event)
 void wxCheckListBox::OnLeftClick(wxMouseEvent& event)
 {
   // clicking on the item selects it, clicking on the checkmark toggles
-  if ( (uint)event.GetX() <= wxOwnerDrawn::GetDefaultMarginWidth() ) {
+  if ( event.GetX() <= wxOwnerDrawn::GetDefaultMarginWidth() ) {
     // # better use LB_ITEMFROMPOINT perhaps?
     uint nItem = ((uint)event.GetY()) / m_nItemHeight;
     if ( nItem < (uint)m_noItems )
